fix(Chat): Correctly initialize socket connection for chat system
Previously the socket connection is initialize in the socket context.
Socket is not initialized at the conversation page and a null error
is shown.

This commit ensure that whenever the chat is loaded the socket
connection is completed first before emitting any services
