+++ src/routes/chat/Chat.jsx
@@ -1,8 +1,9 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import PageContainer from 'components/pageContainer';
-
+import { useSocket } from 'contexts/socketContext';
+import { fetchChatRoomAction, fetchChatListAction } from 'reducers/ChatDux';
 import ChatList from './ChatList';
 import ChatRoom from './ChatRoom/ChatRoom';
 import ChatInput from './ChatInput/ChatInput';
@@ -34,10 +35,22 @@ const ChatContent = () => {
 };
 
 const Chat = () => {
+  const dispatch = useDispatch();
   const { chatRoomId } = useParams();
   const correspondentName = useSelector(
     state => state.chat.chatRoom.length > 0 && state.chat.chatRoom[0].dealerName
   );
+  const { openChatSocket, closeChatSocket } = useSocket();
+  useEffect(() => {
+    openChatSocket();
+    dispatch(fetchChatListAction());
+    if (chatRoomId) {
+      dispatch(fetchChatRoomAction({ chatRoomId }));
+    }
+    return () => {
+      closeChatSocket();
+    };
+  }, [chatRoomId, dispatch, openChatSocket, closeChatSocket]);
 
   return (
     <PageContainer className="chat">
